pyproject.toml
[project]
name = "standup_report"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.4.3",
    "flask>=3.1.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "mypy>=1.19.1",
    "ruff>=0.14.9",
]


